@@ -1,6 +1,6 @@
 <Project Sdk="Microsoft.NET.Sdk">
     <PropertyGroup>
-        <TargetFramework>net7.0</TargetFramework>
+        <TargetFramework>net8.0</TargetFramework>
         <PreserveCompilationContext>false</PreserveCompilationContext>
         <ShouldIncludeNativeSkiaSharp>false</ShouldIncludeNativeSkiaSharp>
         <AssemblyTitle>Artemis.Core</AssemblyTitle>
@@ -36,20 +36,20 @@
     </ItemGroup>
 
     <ItemGroup>
-        <PackageReference Include="DryIoc.dll" Version="5.4.3" />
-        <PackageReference Include="EmbedIO" Version="3.5.2" />
-        <PackageReference Include="HidSharp" Version="2.1.0" />
-        <PackageReference Include="Humanizer.Core" Version="2.14.1" />
-        <PackageReference Include="JetBrains.Annotations" Version="2023.3.0" />
-        <PackageReference Include="McMaster.NETCore.Plugins" Version="1.4.0" />
-        <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
-        <PackageReference Include="RGB.NET.Core" Version="2.0.4-prerelease.1" />
-        <PackageReference Include="RGB.NET.Layout" Version="2.0.4-prerelease.1" />
-        <PackageReference Include="RGB.NET.Presets" Version="2.0.4-prerelease.1" />
-        <PackageReference Include="Serilog.Sinks.Console" Version="5.0.1" />
-        <PackageReference Include="Serilog.Sinks.Debug" Version="2.0.0" />
-        <PackageReference Include="Serilog.Sinks.File" Version="5.0.0" />
-        <PackageReference Include="SkiaSharp" Version="2.88.7" />
+        <PackageReference Include="DryIoc.dll" />
+        <PackageReference Include="EmbedIO" />
+        <PackageReference Include="HidSharp" />
+        <PackageReference Include="Humanizer.Core" />
+        <PackageReference Include="JetBrains.Annotations" />
+        <PackageReference Include="McMaster.NETCore.Plugins" />
+        <PackageReference Include="Newtonsoft.Json" />
+        <PackageReference Include="RGB.NET.Core" />
+        <PackageReference Include="RGB.NET.Layout" />
+        <PackageReference Include="RGB.NET.Presets" />
+        <PackageReference Include="Serilog.Sinks.Console" />
+        <PackageReference Include="Serilog.Sinks.Debug" />
+        <PackageReference Include="Serilog.Sinks.File" />
+        <PackageReference Include="SkiaSharp" />
     </ItemGroup>
 
     <ItemGroup>
