@@ -1,12 +1,14 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
     <PropertyGroup>
-        <TargetFramework>net7.0</TargetFramework>
+        <TargetFramework>net8.0</TargetFramework>
         <PreserveCompilationContext>false</PreserveCompilationContext>
         <Platforms>x64</Platforms>
+        <Nullable>enable</Nullable>
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="LiteDB" Version="5.0.17" />
-        <PackageReference Include="Serilog" Version="3.1.1" />
+        <PackageReference Include="LiteDB" />
+        <PackageReference Include="Serilog" />
+        <PackageReference Include="Newtonsoft.Json" />
     </ItemGroup>
 </Project>
