@@ -1,7 +1,7 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
     <PropertyGroup>
         <OutputType>Library</OutputType>
-        <TargetFramework>net7.0</TargetFramework>
+        <TargetFramework>net8.0</TargetFramework>
         <Nullable>enable</Nullable>
         <OutputPath>bin\</OutputPath>
         <Platforms>x64</Platforms>
@@ -10,17 +10,17 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="Avalonia" Version="11.0.6" />
+        <PackageReference Include="Avalonia" />
         <!--Condition below is needed to remove Avalonia.Diagnostics package from build output in Release configuration.-->
-        <PackageReference Condition="'$(Configuration)' == 'Debug'" Include="Avalonia.Diagnostics" Version="11.0.6" />
-        <PackageReference Include="Avalonia.Controls.ItemsRepeater" Version="11.0.6" />
-        <PackageReference Include="Avalonia.ReactiveUI" Version="11.0.6" />
-        <PackageReference Include="Avalonia.Xaml.Behaviors" Version="11.0.6" />
-        <PackageReference Include="DynamicData" Version="8.3.27" />
-        <PackageReference Include="FluentAvaloniaUI" Version="2.0.5" />
-        <PackageReference Include="Material.Icons.Avalonia" Version="2.1.0" />
-        <PackageReference Include="ReactiveUI" Version="19.5.39" />
-        <PackageReference Include="ReactiveUI.Validation" Version="3.1.7" />
+        <PackageReference Condition="'$(Configuration)' == 'Debug'" Include="Avalonia.Diagnostics" />
+        <PackageReference Include="Avalonia.Controls.ItemsRepeater" />
+        <PackageReference Include="Avalonia.ReactiveUI" />
+        <PackageReference Include="Avalonia.Xaml.Behaviors" />
+        <PackageReference Include="DynamicData" />
+        <PackageReference Include="FluentAvaloniaUI" />
+        <PackageReference Include="Material.Icons.Avalonia" />
+        <PackageReference Include="ReactiveUI" />
+        <PackageReference Include="ReactiveUI.Validation" />
     </ItemGroup>
     <ItemGroup>
         <ProjectReference Include="..\Artemis.Core\Artemis.Core.csproj" />
