@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
     <PropertyGroup>
         <OutputType>WinExe</OutputType>
-        <TargetFramework>net7.0-windows10.0.17763.0</TargetFramework>
+        <TargetFramework>net8.0-windows10.0.17763.0</TargetFramework>
         <Nullable>enable</Nullable>
         <Platforms>x64</Platforms>
         <OutputPath>bin</OutputPath>
@@ -21,13 +21,12 @@
         </None>
     </ItemGroup>
     <ItemGroup>
-        <PackageReference Include="Avalonia.Win32" Version="11.0.6" />
-        <PackageReference Include="Microsoft.Toolkit.Uwp.Notifications" Version="7.1.3" />
-        <PackageReference Include="Microsoft.Win32" Version="2.0.1" />
-        <!-- Note: Do NOT upgrade this compatibility package to 8.X before updating to net8, it WILL break -->
-        <PackageReference Include="Microsoft.Windows.Compatibility" Version="7.0.5" />
-        <PackageReference Include="RawInput.Sharp" Version="0.1.3" />
-        <PackageReference Include="SkiaSharp.Vulkan.SharpVk" Version="2.88.7" />
+        <PackageReference Include="Avalonia.Win32" />
+        <PackageReference Include="Microsoft.Toolkit.Uwp.Notifications" />
+        <PackageReference Include="Microsoft.Win32" />
+        <PackageReference Include="Microsoft.Windows.Compatibility" />
+        <PackageReference Include="RawInput.Sharp" />
+        <PackageReference Include="SkiaSharp.Vulkan.SharpVk" />
     </ItemGroup>
     <ItemGroup>
         <ProjectReference Include="..\Artemis.Core\Artemis.Core.csproj" />
