@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
     <PropertyGroup>
         <OutputType>Library</OutputType>
-        <TargetFramework>net7.0</TargetFramework>
+        <TargetFramework>net8.0</TargetFramework>
         <Nullable>enable</Nullable>
         <OutputPath>bin/</OutputPath>
         <Platforms>x64</Platforms>
@@ -10,26 +10,25 @@
     <ItemGroup>
         <ProjectReference Include="..\Artemis.Core\Artemis.Core.csproj" />
         <ProjectReference Include="..\Artemis.UI.Shared\Artemis.UI.Shared.csproj" />
-        <ProjectReference Include="..\Artemis.VisualScripting\Artemis.VisualScripting.csproj" />
         <ProjectReference Include="..\Artemis.WebClient.Updating\Artemis.WebClient.Updating.csproj" />
         <ProjectReference Include="..\Artemis.WebClient.Workshop\Artemis.WebClient.Workshop.csproj" />
     </ItemGroup>
 
     <ItemGroup>
-        <PackageReference Include="AsyncImageLoader.Avalonia" Version="3.2.1" />
-        <PackageReference Include="Avalonia.AvaloniaEdit" Version="11.0.6" />
-        <PackageReference Include="Avalonia.Controls.PanAndZoom" Version="11.0.0.2" />
-        <PackageReference Include="Avalonia.Desktop" Version="11.0.6" />
-        <PackageReference Include="Avalonia.Skia.Lottie" Version="11.0.0" />
-        <PackageReference Include="AvaloniaEdit.TextMate" Version="11.0.6" />
-        <PackageReference Include="Markdown.Avalonia.Tight" Version="11.0.2" />
-        <PackageReference Include="Octopus.Octodiff" Version="2.0.544" />
-        <PackageReference Include="PropertyChanged.SourceGenerator" Version="1.1.0">
+        <PackageReference Include="AsyncImageLoader.Avalonia" />
+        <PackageReference Include="Avalonia.AvaloniaEdit" />
+        <PackageReference Include="Avalonia.Controls.PanAndZoom" />
+        <PackageReference Include="Avalonia.Desktop" />
+        <PackageReference Include="Avalonia.Skia.Lottie" />
+        <PackageReference Include="AvaloniaEdit.TextMate" />
+        <PackageReference Include="Markdown.Avalonia.Tight" />
+        <PackageReference Include="Octopus.Octodiff" />
+        <PackageReference Include="PropertyChanged.SourceGenerator">
           <PrivateAssets>all</PrivateAssets>
           <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
         </PackageReference>
-        <PackageReference Include="Splat.DryIoc" Version="14.8.12" />
-        <PackageReference Include="TextMateSharp.Grammars" Version="1.0.56" />
+        <PackageReference Include="Splat.DryIoc" />
+        <PackageReference Include="TextMateSharp.Grammars" />
     </ItemGroup>
 
     <ItemGroup>
