@@ -1,17 +1,17 @@
 <Project Sdk="Microsoft.NET.Sdk">
     <PropertyGroup>
-        <TargetFramework>net7.0</TargetFramework>
+        <TargetFramework>net8.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
         <Platforms>x64</Platforms>
     </PropertyGroup>
     
     <ItemGroup>
-      <PackageReference Include="DryIoc.Microsoft.DependencyInjection" Version="6.2.0" />
-      <PackageReference Include="IdentityModel" Version="6.2.0" />
-      <PackageReference Include="Microsoft.Extensions.Http" Version="8.0.0" />
-      <PackageReference Include="StrawberryShake.Server" Version="13.8.1" />
-      <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="7.2.0" />
+      <PackageReference Include="DryIoc.Microsoft.DependencyInjection" />
+      <PackageReference Include="IdentityModel" />
+      <PackageReference Include="Microsoft.Extensions.Http" />
+      <PackageReference Include="StrawberryShake.Server" />
+      <PackageReference Include="System.IdentityModel.Tokens.Jwt" />
     </ItemGroup>
 
     <ItemGroup>
