@@ -8,8 +8,10 @@ using System.Text;
 using System.Threading.Tasks;
 using Artemis.Core.Modules;
 using Artemis.Storage.Entities.Profile;
+using Artemis.Storage.Migrations;
 using Artemis.Storage.Repositories.Interfaces;
 using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
 using Serilog;
 using SkiaSharp;
 
@@ -24,9 +26,10 @@ internal class ProfileService : IProfileService
     private readonly List<ArtemisKeyboardKeyEventArgs> _pendingKeyboardEvents = new();
     private readonly List<ProfileCategory> _profileCategories;
     private readonly IProfileRepository _profileRepository;
+    private readonly List<IProfileMigration> _profileMigrators;
     private readonly List<Exception> _renderExceptions = new();
     private readonly List<Exception> _updateExceptions = new();
-    
+
     private DateTime _lastRenderExceptionLog;
     private DateTime _lastUpdateExceptionLog;
 
@@ -35,13 +38,15 @@ internal class ProfileService : IProfileService
         IPluginManagementService pluginManagementService,
         IInputService inputService,
         IDeviceService deviceService,
-        IProfileRepository profileRepository)
+        IProfileRepository profileRepository,
+        List<IProfileMigration> profileMigrators)
     {
         _logger = logger;
         _profileCategoryRepository = profileCategoryRepository;
         _pluginManagementService = pluginManagementService;
         _deviceService = deviceService;
         _profileRepository = profileRepository;
+        _profileMigrators = profileMigrators;
         _profileCategories = new List<ProfileCategory>(_profileCategoryRepository.GetAll().Select(c => new ProfileCategory(c)).OrderBy(c => c.Order));
 
         _deviceService.LedsChanged += DeviceServiceOnLedsChanged;
@@ -58,7 +63,7 @@ internal class ProfileService : IProfileService
     public ProfileConfiguration? FocusProfile { get; set; }
     public ProfileElement? FocusProfileElement { get; set; }
     public bool UpdateFocusProfile { get; set; }
-    
+
     public bool ProfileRenderingDisabled { get; set; }
 
     /// <inheritdoc />
@@ -221,7 +226,7 @@ internal class ProfileService : IProfileService
             return profileConfiguration.Profile;
         }
 
-        ProfileEntity profileEntity;
+        ProfileEntity? profileEntity;
         try
         {
             profileEntity = _profileRepository.Get(profileConfiguration.Entity.ProfileId);
@@ -280,7 +285,7 @@ internal class ProfileService : IProfileService
     {
         DeactivateProfile(profileConfiguration);
 
-        ProfileEntity profileEntity = _profileRepository.Get(profileConfiguration.Entity.ProfileId);
+        ProfileEntity? profileEntity = _profileRepository.Get(profileConfiguration.Entity.ProfileId);
         if (profileEntity == null)
             return;
 
@@ -353,7 +358,7 @@ internal class ProfileService : IProfileService
 
         DeactivateProfile(profileConfiguration);
         SaveProfileCategory(profileConfiguration.Category);
-        ProfileEntity profileEntity = _profileRepository.Get(profileConfiguration.Entity.ProfileId);
+        ProfileEntity? profileEntity = _profileRepository.Get(profileConfiguration.Entity.ProfileId);
         if (profileEntity != null)
             _profileRepository.Remove(profileEntity);
 
@@ -461,7 +466,12 @@ internal class ProfileService : IProfileService
 
         await using Stream profileStream = profileEntry.Open();
         using StreamReader profileReader = new(profileStream);
-        ProfileEntity? profileEntity = JsonConvert.DeserializeObject<ProfileEntity>(await profileReader.ReadToEndAsync(), IProfileService.ExportSettings);
+        JObject? profileJson = JsonConvert.DeserializeObject<JObject>(await profileReader.ReadToEndAsync(), IProfileService.ExportSettings);
+
+        // Before deserializing, apply any pending migrations
+        MigrateProfile(configurationEntity, profileJson);
+
+        ProfileEntity? profileEntity = profileJson?.ToObject<ProfileEntity>(JsonSerializer.Create(IProfileService.ExportSettings));
         if (profileEntity == null)
             throw new ArtemisCoreException("Could not import profile, failed to deserialize profile.json");
 
@@ -514,10 +524,10 @@ internal class ProfileService : IProfileService
     public async Task<ProfileConfiguration> OverwriteProfile(MemoryStream archiveStream, ProfileConfiguration profileConfiguration)
     {
         ProfileConfiguration imported = await ImportProfile(archiveStream, profileConfiguration.Category, true, true, null, profileConfiguration.Order + 1);
-        
+
         DeleteProfile(profileConfiguration);
         SaveProfileCategory(imported.Category);
-        
+
         return imported;
     }
 
@@ -542,6 +552,21 @@ internal class ProfileService : IProfileService
         lock (_profileCategories)
         {
             _pendingKeyboardEvents.Add(e);
+        }
+    }
+
+    private void MigrateProfile(ProfileConfigurationEntity configurationEntity, JObject? profileJson)
+    {
+        if (profileJson == null)
+            return;
+
+        foreach (IProfileMigration profileMigrator in _profileMigrators.OrderBy(m => m.Version))
+        {
+            if (profileMigrator.Version <= configurationEntity.Version)
+                continue;
+
+            profileMigrator.Migrate(profileJson);
+            configurationEntity.Version = profileMigrator.Version;
         }
     }
 
