@@ -11,7 +11,7 @@ public class QueuedActionEntity
     }
 
     public Guid Id { get; set; }
-    public string Type { get; set; }
+    public string Type { get; set; } = string.Empty;
     public DateTimeOffset CreatedAt { get; set; }
 
     public Dictionary<string, object> Parameters { get; set; }
