@@ -6,7 +6,7 @@ public class ScriptConfigurationEntity
 {
     public Guid Id { get; set; }
 
-    public string Name { get; set; }
-    public string ScriptingProviderId { get; set; }
-    public string ScriptContent { get; set; }
+    public string Name { get; set; } = string.Empty;
+    public string ScriptingProviderId { get; set; } = string.Empty;
+    public string? ScriptContent { get; set; }
 }
