@@ -10,6 +10,6 @@ public class PluginSettingEntity
     public Guid Id { get; set; }
     public Guid PluginGuid { get; set; }
 
-    public string Name { get; set; }
-    public string Value { get; set; }
+    public string Name { get; set; } = string.Empty;
+    public string Value { get; set; } = string.Empty;
 }
