@@ -8,6 +8,6 @@ public class EventConditionEntity : IConditionEntity
     public int TriggerMode { get; set; }
     public int OverlapMode { get; set; }
     public int ToggleOffMode { get; set; }
-    public DataModelPathEntity EventPath { get; set; }
-    public NodeScriptEntity Script { get; set; }
+    public DataModelPathEntity? EventPath { get; set; }
+    public NodeScriptEntity? Script { get; set; }
 }
