@@ -1,5 +1,3 @@
 ﻿namespace Artemis.Storage.Entities.Profile.Abstract;
 
-public interface IConditionEntity
-{
-}
+public interface IConditionEntity;
