@@ -2,8 +2,8 @@
 
 public class LayerBrushEntity
 {
-    public string ProviderId { get; set; }
-    public string BrushType { get; set; }
+    public string ProviderId { get; set; } = string.Empty;
+    public string BrushType { get; set; } = string.Empty;
 
-    public PropertyGroupEntity PropertyGroup { get; set; }
+    public PropertyGroupEntity? PropertyGroup { get; set; }
 }
