@@ -2,11 +2,11 @@
 
 public class LayerEffectEntity
 {
-    public string ProviderId { get; set; }
-    public string EffectType { get; set; }
-    public string Name { get; set; }
+    public string ProviderId { get; set; } = string.Empty;
+    public string EffectType { get; set; } = string.Empty;
+    public string Name { get; set; } = string.Empty;
     public bool HasBeenRenamed { get; set; }
     public int Order { get; set; }
 
-    public PropertyGroupEntity PropertyGroup { get; set; }
+    public PropertyGroupEntity? PropertyGroup { get; set; }
 }
