@@ -12,22 +12,21 @@ public class LayerEntity : RenderElementEntity
     {
         Leds = new List<LedEntity>();
         AdaptionHints = new List<IAdaptionHintEntity>();
-        LayerEffects = new List<LayerEffectEntity>();
     }
 
     public int Order { get; set; }
-    public string Name { get; set; }
+    public string? Name { get; set; }
     public bool Suspended { get; set; }
 
     public List<LedEntity> Leds { get; set; }
     public List<IAdaptionHintEntity> AdaptionHints { get; set; }
 
-    public PropertyGroupEntity GeneralPropertyGroup { get; set; }
-    public PropertyGroupEntity TransformPropertyGroup { get; set; }
-    public LayerBrushEntity LayerBrush { get; set; }
+    public PropertyGroupEntity? GeneralPropertyGroup { get; set; }
+    public PropertyGroupEntity? TransformPropertyGroup { get; set; }
+    public LayerBrushEntity? LayerBrush { get; set; }
 
     [BsonRef("ProfileEntity")]
-    public ProfileEntity Profile { get; set; }
+    public ProfileEntity Profile { get; set; } = null!;
 
     public Guid ProfileId { get; set; }
 }
