@@ -10,8 +10,8 @@ public class NodeScriptEntity
         Connections = new List<NodeConnectionEntity>();
     }
 
-    public string Name { get; set; }
-    public string Description { get; set; }
+    public string Name { get; set; } = string.Empty;
+    public string Description { get; set; } = string.Empty;
 
     public List<NodeEntity> Nodes { get; set; }
     public List<NodeConnectionEntity> Connections { get; set; }
