@@ -10,13 +10,12 @@ public class EntryEntity
     public long EntryId { get; set; }
     public int EntryType { get; set; }
     
-    public string Author { get; set; } 
+    public string Author { get; set; }  = string.Empty;
     public string Name { get; set; } = string.Empty;
-    public string Summary { get; set; } = string.Empty;
 
     public long ReleaseId { get; set; }
-    public string ReleaseVersion { get; set; }
+    public string ReleaseVersion { get; set; } = string.Empty;
     public DateTimeOffset InstalledAt { get; set; }
 
-    public Dictionary<string,object> Metadata { get; set; }
+    public Dictionary<string,object>? Metadata { get; set; }
 }
