@@ -6,12 +6,12 @@ namespace Artemis.Storage.Repositories.Interfaces;
 public interface IPluginRepository : IRepository
 {
     void AddPlugin(PluginEntity pluginEntity);
-    PluginEntity GetPluginByGuid(Guid pluginGuid);
+    PluginEntity? GetPluginByGuid(Guid pluginGuid);
     void SavePlugin(PluginEntity pluginEntity);
 
     void AddSetting(PluginSettingEntity pluginSettingEntity);
-    PluginSettingEntity GetSettingByGuid(Guid pluginGuid);
-    PluginSettingEntity GetSettingByNameAndGuid(string name, Guid pluginGuid);
+    PluginSettingEntity? GetSettingByGuid(Guid pluginGuid);
+    PluginSettingEntity? GetSettingByNameAndGuid(string name, Guid pluginGuid);
     void SaveSetting(PluginSettingEntity pluginSettingEntity);
     void RemoveSettings(Guid pluginGuid);
 }
