@@ -9,6 +9,6 @@ public interface IProfileRepository : IRepository
     void Add(ProfileEntity profileEntity);
     void Remove(ProfileEntity profileEntity);
     List<ProfileEntity> GetAll();
-    ProfileEntity Get(Guid id);
+    ProfileEntity? Get(Guid id);
     void Save(ProfileEntity profileEntity);
 }
