@@ -1,5 +1,3 @@
 ﻿namespace Artemis.Storage.Repositories.Interfaces;
 
-public interface IRepository
-{
-}
+public interface IRepository;
